virt-install-bionic: Add support for kickstart
Add the ability to do a complete unguided install of a LDAP and NFS
enabled VM on the Eideticom network. Note for LDAP you need to inject
some files that are not in this repo and run:

NAME=milan KS=eideticom-vm.ks \
  INITRD_INJECT="--initrd-inject /home/users/sbates/ldap/ldap.conf" \
  "--initrd-inject /home/users/sbates/ldap/ldap.secret" \
  ./virt-install-bionic

This only works for Ubuntu installers (18.04 tested). RHEL derivatives
use a slightly different syntax so they won't work.

Signed-off-by: Stephen Bates <sbates@raithlin.com>