virt-install-ubuntu: Add support for Focal VM creation
Update virt-install-bionic and rename it virt-install-ubuntu to
support both Bionic and Focal creation via a RELEASE option on the
command line (default is Focal, use 'RELEASE=bionic' for Bionic).

Canonical moved to preferring cloud-images and cloud-init from Focal
on so we use cloud-init to tune the default Focal disk image to our
needs. This includes:
  1. Set hostname to NAME.
  2. Create user ubuntu with password 'change' with full sudo.
  3. Setup DHCP networking via libvirt NAT.
  4. Disable root account.

Note there are a couple of fixups I would still like to do to this
around how we inject the disk images into libvirt pools.

Signed-off-by: Stephen Bates <sbates@raithlin.com>