gen-image: Add a script to generate a cloud-init based image
Cloud-init provides a very nice way to setup a quick image file for a
VM based on Ubuntu releases for amd64 and arm64. This script automates
this process and is based on a script developed by Martin. This script
generates two images and when booted runs a cloud-init process to
setup the first user.

Fixes #10.