virt-install-ubuntu: Update to make more flexible
We update this script for the cloud-init cases to include the
following:
  1. Allow for non-ubuntu username.
  2. Use plaintext password.
  3. Allow insertion of a ssh public key for remote access.
  4. Improved network setup by doing a more glob based match on the
     interface name. This helps when working across different ARCH.
  5. Better trapping and cleanup of scripts.
  6. Support for non-amd64 ARCH.
  7. Add ability to not connect to console via NOAUTOCONSOLE=true
     environmaent variable.
  8. Update the script to use the virt-install --cloud-init rather
     than adding the ISO ourself.
  9. Fix an issue where the VM was not shutdown at the end of the
     install. Now cloud-init makes sure the VM powers down.
  10.Remove the .cfg from the network cloud-init configuration to be
     consistent with the user filename.

Note that there are some review comments to improve how the
SSH_KEY_FILE input is handled and to add a mode where the number of
vcpus can be provided by the user. We open new issues for these and
will address them later.

An example of this script might be:

NAME=stephen-$(date '+%s') SSH_KEY_FILE=~/.ssh/id_rsa.pub USERNAME=batesste ./virt-install-ubuntu