rebase: got arm working inside my VM at last
Note: Some sort of issue with the rootfs not booting unless we add a
uefi. But is there not already one on the cloud image??

Note: There does seem to be a schema error in my cloud-config file. I
think it is related to the ssh authorized key. Investigate. Still
works as expected.

Useful - https://gist.github.com/george-hawkins/16ee37063213f348a17717a7007d2c79