qemu/run-vm: Add dynamic support for NVMe SSD emulation
We improve how the user can ask for emulated NVMe SSDs when invoking
the run-vm script. Now the user can pass in true, a positive number or
a literal to create emulated NVMe SSD(s) to attach to the VM. See the
header of the run-vm script for more details.