gen_vm: Add a USER_ID argument to set uid
By default the uid for the user is 1000. However on some systems we
want to request a different uid because the host user has a differnt
uid and we want to pass the filesystem in and avoid a need to map from
one uid (in guest) to another (on host).

Signed-off-by: Stephen Bates <sbates@raithlin.com>