run-vm: Update for pci-mmio-bridge
Update our run-vm script so we can call the pci-mmio-bridge when it is
available to us in QEMU. Currently it only exists as an external fork
of upstream QEMU. However it is very useful for a range of differnet
things.

We also ensure that emulated NVMe disable shadow doorbells because
these do not work with the pci-mmio-bridge method of doorbell ringing.

[1]: https://github.com/sbates130272/qemu

Signed-off-by: Stephen Bates <sbates@raithlin.com>